+++ epoch1d/src/io/calc_df.F90
@@ -759,7 +759,7 @@ CONTAINS
     REAL(num) :: root, fac, gamma, cf2
     REAL(num) :: delta_x, part_vy, part_vz
     INTEGER :: ispecies, ix, dcellx
-    INTEGER(KIND=8) :: ipart
+    INTEGER(i8) :: ipart
     TYPE(particle), POINTER :: current, next
     INTEGER :: spec_start, spec_end
     LOGICAL :: spec_sum
